backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.0.8
openpyxl==3.1.5
gunicorn==23.0.0
python-dotenv==1.0.1
